perf(core): improve hash table lookup performance for small string keys such as column names and symbols

+++ benchmarks/src/main/java/org/questdb/GroupByLongHashSetBenchmark.java
@@ -78,7 +78,7 @@ public class GroupByLongHashSetBenchmark {
     @Benchmark
     public void testGroupByLongHashSet() {
         long value = rnd.nextLong(size);
-        int index = groupByLongHashSet.of(ptr).keyIndex(value);
+        long index = groupByLongHashSet.of(ptr).keyIndex(value);
         if (index >= 0) {
             groupByLongHashSet.addAt(index, value);
             ptr = groupByLongHashSet.ptr();
